Add Active Show Panel
This commit adds the basic functionality required for a user to select a
schedule item and see more data about that show. It does so by creating
a new parent component for the ScheduleList component, SchedulePanel, and
rendering in that parent both the ScheduleList and the new
ShowDetailPanel components. The active show id is set by passing the
setActiveShowId function from the SchedulePanel down to the ScheduleItem
component, and calling it on click. Rather than passing the show object
itself as state, the id is picked and used as a filter for the shows
array in-memory.