Add Navbar
This commit adds a new component, Navbar, to the app layout, meaning it
will appear on every page. Currently, the navbar only provides the
stream embed and links to the Home and Schedule page, with minimal
styling. As new pages are added, they will be appended to the list of
links. Further development on the stream is currently on halt until the
API is prepped to receive data from the spinitron's meta-data push
messaging channel. When this is resolved, we will be to display the
latest data from spinitron without having to worry about over-querying
their servers.