Add polling for stream
This commit adds a polling system to the stream component,
which queries our proxy api every five seconds for the latest stream
data. It does this via a custom react hook, defined in the new api.ts
folder. This was directly referenced from the pdc front-end repository,
thanks to @slifty and @reefdog for their excellent documentation.
It also adds the @wnyu/spinitron-sdk as a core module in the eslintrc
file, as it was throwing an error for some reason