Add embedded sanity studio
This commit adds an embedded sanity studio to the application,
accessible at `/admin`.  Sanity has been chosen as the CMS for this
project namely because of it's nextjs integration, and direct
recommendation from peers. In order to accommodate the embedded studio,
the router has been reorganized slightly, taking advantage of nextjs
routing groups [1] to isolate the the page layouts from the studio.

[1] https://nextjs.org/docs/app/building-your-application/routing/route-groups