Introduce Dropdown Panel
This sizable commit introduces a dropdown panel to the mobile layout of
the application which shows the current show and playlist.  In doing
this, it also refactors a significant amount of the codebase,
principally in introducing the new 'Providers', which provide global
context to the application. A design concern was having two Show Panels
polling the API separately for up to date playlist data (the previous
ActiveShowPanel, and now the dropdown). Since they were both querying
the same endpoint, we want something that can poll the endpoint ONCE,
and distribute the data to the components as needed, minimizing the
overall number of requests to the API. As such, we introduce
CurrentDataProvider, which wraps the entire application, and follows the
context/provider pattern laid out in the react docs.

This may have ramifications in terms of SSR optimizations, but it's
benefits in minimizing data fetching currently outweigh
this. Furthermore, many styles have been refactored and components
redesigned to make room for this new panel and way of fetching data. We
are still using separate fetches for the upcoming show information and
the metadata endpoint (as it is reliably the most up-to-date
information from Spinitron).

Issue #69-create-drop-down-live-panel