Add show tiles to schedule page
This commit adds an alternative method of viewing the
shows in the schedule, as a two-columned grid of show images
which is selectable via the ShowDetailPanel in desktop mode.
It does not currently function properly for mobile and will need to
be tweaked further, once we get a proper mobile schedule layout
mockup.

Issue #72-add-show-tiles-to-schedule-page