Add Announcements Slideshow
This commit adds a slideshow to the homepage that is populated by Sanity.
It adds the swiper.js library to the project, which is the most popularly used
open-source slideshow library for javascript. It also rearranges the layout of
the homepage somewhat, to fit both the currentPlaylist panel and carousel together.

Issue #71-add-announcements-slideshow