Style Schedule Filter
This commit follows the pattern of the navbarmenu component and
pulls the Schedule search bar funcitonality into it's own component,
principally to control rendering based on screen size (using media queries to hide or display mobile or desktop layouts). This seems to be an effective approach for decisioning which layout is rendered at the component level.

Issue #101-style-schedule-filter-mobile