Add Mobile Playlist Page
This commit adds a dynamic route and page for playlists, the recorded
instance of shows, which display the show name, host name, date of play,
and tracks played from the playlist. It only implements a mobile layout.

This is the first time we are fetching data from the api on a server-only component, and presents an opportunity to cache data that does not need to be updated regularly. The approach outlined in the nextjs fetch docs was used here: https://nextjs.org/docs/app/building-your-application/data-fetching/fetching#fetch-api. We note that moving forward a more adept data fetching and caching approach may be adopted.