Add Video Card type in sanity
This commit adds the new video card type into sanity, which contains
only a link to a video (which we will need to validate on our own end),
and adds the necessary types and queries to access this data on the home
page. This is also the first time we're using the required validator for
a sanity schema, but reasonably we should be doing this for more
schemas. I'll make an issue for this.