Refactor Show Page into schedule page
This lengthy commit refactors the show page into the schedule page, for
the sake of slickly combining the two interfaces for a desktop
layout. Effectively, we wanted to load the show page in the schedule
page on the side panel, but in order to do that without reusing a bunch
of code, we needed a way to render them together. To preserve
server-side rendering, we have now also extracted the activeShowId
parameter from a reactive element to a url parameter, woohoo. Following
the same pattern of defining a mobile and desktop component and hiding
using media queries, we have redesigned the schedule panel component to
now contain both show detail (which is now a show page, effectively) and
the schedule. I'm the best at design patterns