Refactor show page for desktop
This horendously ugly commit changes many spacing related things in the
app in pursuit of a more sensable spacing layout for everything, but
also mostly just implements a workable desktop page interface for
shows. Show pages now are the same as schedule pages, which take in a
show id and use that to display the active show, woohoo.